DEV: Make `store` an actual service
